WIP knocking down the errors, Trace logic updated

+++ include/battle_util.h
@@ -99,10 +99,10 @@ u8 TryWeatherFormChange(u8 battlerId);
 bool32 TryChangeBattleWeather(u8 battler, u32 weatherEnumId, bool32 viaAbility);
 u8 AbilityBattleEffects(u8 caseID, u8 battlerId, u16 ability, u8 special, u16 moveArg);
 u16 *GetBattlerAbilities(u8 battlerId);
-u32 IsAbilityOnSide(u32 battlerId, u32 ability);
-u32 IsAbilityOnOpposingSide(u32 battlerId, u32 ability);
-u32 IsAbilityOnField(u32 ability);
-u32 IsAbilityOnFieldExcept(u32 battlerId, u32 ability);
+u32 IsAbilityOnSide(u32 battlerId, u16 ability);
+u32 IsAbilityOnOpposingSide(u32 battlerId, u16 ability);
+u32 IsAbilityOnField(u16 ability);
+u32 IsAbilityOnFieldExcept(u32 battlerId, u16 ability);
 u32 IsAbilityPreventingEscape(u32 battlerId);
 bool32 IsBattlerProtected(u8 battlerId, u16 move);
 bool32 CanBattlerEscape(u32 battlerId); // no ability check
